fix: executerのヒアドキュメントに関連する箇所を別ファイルに
