+++ Makefile
@@ -14,7 +14,7 @@ BUILTIN_SRCS = $(addprefix srcs/builtin/, \
 		cd.c echo.c env.c exit.c export.c pwd.c unset.c \
 		)
 EXECUTER_SRCS = $(addprefix srcs/executer/, \
-		builtin.c executer.c fd_getter.c get_fullpath.c heredoc.c\
+		builtin.c executer.c fd_getter.c get_fullpath.c heredoc.c signal.c \
 		)
 EXPANSION_SRCS = $(addprefix srcs/expansion/, \
 		expansion.c \
